Rename the 3D debug-deserialize example
